Add files via upload
Spam is sent by both reputable organisations and lesser companies. When spam is sent by
reputable companies it is sometimes referred to as Main sleaze. Main sleaze makes up
approximately 3% of the spam sent over the internet. The problem with main sleaze is that it is
generally mixed in with mail that the recipients asked for, and it is difficult to tell the difference
using traditional mail filters. As a result, if a mail system filters out all the mail from a
mainstream, they will get complaints from the people who signed up.
It was estimated in 2009 that spam cost businesses around US$130 billion. As the scale of the
spam problem has grown, ISP s and the public have turned to government for relief from spam,
which has failed to materialise.
Spam is also a medium for fraudsters to scam users into entering personal information on fake
Web sites using emails forged to look like they are from banks or other organisations, such
as PayPal. This is known as phishing. Targeted phishing, where known information about the
recipient is used to create forged emails, is known as spear-phishing.

This Repository contains the full documentation of the project as well.